requirements.txt
flask
flask-mail
pillow
pyyaml
